VS2017: added GzFile.cpp, Logger.cpp, Signal.cpp

+++ src/core/GzFile.cpp
@@ -334,7 +334,7 @@ size_t GzFile::fread(void * ptr, size_t size, size_t nmemb)
 {
 	if (ff==NULL) throw FileNotOpenException();
 	if (ptr==NULL) throw IllegalArgumentException();
-	int by=::gzread((gzFile)ff,ptr,size*nmemb);
+	int by=::gzread((gzFile)ff,ptr,(unsigned int)(size*nmemb));
 	if (by>0) return by;
 	if (by==0) throw ppl7::EndOfFileException();
 	int err=0;
@@ -348,7 +348,7 @@ char * GzFile::fgets (char *buffer, size_t num)
 	if (buffer==NULL) throw IllegalArgumentException();
 	//int suberr;
 	char *res;
-	res=::gzgets((gzFile)ff,buffer, num);
+	res=::gzgets((gzFile)ff,buffer, (int)num);
 	if (res==NULL) {
 		//suberr=::ferror((FILE*)ff);
 		if (gzeof((gzFile)ff)) throw ppl7::EndOfFileException();
